requirements.txt
elevenlabs==2.34.0
python-dotenv==1.2.1
